Improved fetching next characters
Added concatArrayEagerDelayError to repository when fetching data, this is used for when the api fails, I'll use the data from the db.
Added a viewstate checker for the datastate error, to show characters if there were characters saved in the database.
Added todo to remove previous and next button
Removed a few unused methods
Added loop to fetching characters to allow me to continuously fetch characters